article,comments revise

+++ final-pjt-back/community/views.py
@@ -46,33 +46,48 @@ def article_detail(request, article_pk):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     elif request.method == 'PUT':
-        print('here')
         serializer = ArticleSerializer(article, data=request.data)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
 
 
-@api_view(['GET'])
-def comment_list(request):
+@api_view(['GET', 'POST'])
+def comment_list(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'GET':
         comments = get_list_or_404(Comment)
         serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
+    
+    elif request.method == 'POST': # 댓글
+        serializer = CommentSerializer(data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            # serializer.save()
+            serializer.save(user=request.user, article_id=article.pk)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
 @api_view(['GET', 'DELETE', 'PUT'])
-def comment_detail(request, comment_pk):
+def comment_detail(request, article_pk, comment_pk):
     # comment = Comment.objects.get(pk=comment_pk)
     comment = get_object_or_404(Comment, pk=comment_pk)
-
+    article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'GET':
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
 
     elif request.method == 'DELETE':
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        writer = comment.user
+        if request.user == writer: #요청유저와 글쓴 유저가 같다면
+            comment.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            bad = True
+            context = {
+                bad: 'bad'
+            }
+            return Response(context)
 
     elif request.method == 'PUT':
         serializer = CommentSerializer(comment, data=request.data)
@@ -80,15 +95,13 @@ def comment_detail(request, comment_pk):
             serializer.save()
             return Response(serializer.data)
 
-    
-
-
-@api_view(['POST'])
-def comment_create(request, article_pk):
-    # article = Article.objects.get(pk=article_pk)
-    article = get_object_or_404(Article, pk=article_pk)
-    serializer = CommentSerializer(data=request.data)
-    if serializer.is_valid(raise_exception=True):
-        serializer.save(article=article)
-        return Response(serializer.data, status=status.HTTP_201_CREATED)
+# @api_view(['POST'])
+# @permission_classes([IsAuthenticated]) # 인증된 경우만 권한을 준다.
+# def comment_create(request, article_pk):
+#     # article = Article.objects.get(pk=article_pk)
+#     comment = get_object_or_404(Comment, pk=article_pk)
+#     serializer = CommentSerializer(data=request.data)
+#     if serializer.is_valid(raise_exception=True):
+#         serializer.save(article=article)
+#         return Response(serializer.data, status=status.HTTP_201_CREATED)
